+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.138.1](https://github.com/heywood8/money-tracker/compare/penny-v0.138.0...penny-v0.138.1) (2026-06-18)
+
+
+### Bug Fixes
+
+* mark installed version as latest when newer release has no APK ([#1006](https://github.com/heywood8/money-tracker/issues/1006)) ([c1e05a9](https://github.com/heywood8/money-tracker/commit/c1e05a9efeaf5110628b48bbed6887528a5fd0b4))
+
+## [0.138.0](https://github.com/heywood8/money-tracker/compare/penny-v0.137.4...penny-v0.138.0) (2026-06-18)
+
+
+### Features
+
+* show CI build progress when a release APK is not yet available ([#1003](https://github.com/heywood8/money-tracker/issues/1003)) ([c862e20](https://github.com/heywood8/money-tracker/commit/c862e20252e7e87a29d6bf8bf14950a60a4e5dfc))
+
 ## [0.137.4](https://github.com/heywood8/money-tracker/compare/penny-v0.137.3...penny-v0.137.4) (2026-06-18)